Finished initial implementation of voice controlled weather data fetching and displaying
The program is now able to find the weather of any city/province/country in the world and return the weather data for the specified time, such as "next monday". Due to the API's restraints, it can only do so for the next 48 hours hourly, and the next 8 days for daily data. The location detection has been changed to be handled by wit instead of openweatherapp since it is smarter.